Initial Functionality Commit
Only works by using the direct Target URL ( the folder path that the
zoomify image refers to - NOT the actual URL of the page that the
zoomify viewer is on) at this iteration.  Batch handling of direct
target URLs is the next planned update, and then web viewed URL
downloading after.